Reparse 2014 precinct files from SOS sources
All the 2014 precinct files were re-parsed with the command:

```bash
% ruby src/parse-csv.rb -d 20141104 -o 2014 -t general \
  ../openelections-sources-ks/2014/2014_general_governor_official_results.csv.gz \
  ../openelections-sources-ks/2014/2014_general_ks_house_official_results.csv.gz \
  ../openelections-sources-ks/2014/2014_general_us_house_official_results.csv.gz \
  ../openelections-sources-ks/2014/2014-General-Election---*csv.gz
```

Note that the new "totals" file is gleened from the original precinct files, in
order to preserve the Voter stats per-county. Instead of including the totals
within each county file (which mixed the content model), all the counties
that had totals are grouped in a single file.